draw hit objects on play area

+++ src/singleton/map_manager.h
@@ -34,7 +34,7 @@ class MapManager : public Node {
 
     String extract_audio_extension(String file_path);
 
-    float circle_size_to_pixel_for_1080(float circle_size);
+    int64_t circle_size_to_pixel_for_1080(float circle_size);
 
    public:
     static MapManager* get_singleton(Node* node);
@@ -68,7 +68,7 @@ class MapManager : public Node {
     int64_t get_max_placeable_y_coordinate();
     int64_t approach_rate_to_ms(float approach_rate);
     int64_t get_fade_out_time();
-    float circle_size_to_pixel(float circle_size);
+    int64_t circle_size_to_pixel(float circle_size);
 };
 
 #endif
